refactor(version_tag): 🎨 Seperate VersionTag struct and VersionCalculator struct for different purposes
VersionTag doucments a VersionTag and Object ID for that tag.  The latest function generates
the tag for the most recent version based on the defined prefix.
The version Calculator struct examines the commits since the VersionTag
to  calculate the next version number  and documents the level
and the conventional commits discovered.