feat(calculator): has_required function
Takes files_required as option.
Checks list of files required against the files found.
Returns an error if one or more is not found.
Error includes a list of the files that have not been found.
Returns an error if the files list has not been prepared.
If no required files have been specified returns success.