feat(cli): required_level to enforce
long form only --require-level option to speficy the level of
commits that will require enfocement of required files.
Calculate the level by checking for breaking and the counts
of feature and fix commits.
Compare the level calculcated to that requested and only check for the
required files if the threshold has been met.